# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  handler.ErrorResponseBody:
    description: Error response body
    properties:
      error:
        example: something went wrong
        type: string
    type: object
...
  handler.RoleResponse:
    description: Role data
    properties:
      abilities:
        example:
        - eliminate
        items:
          type: string
        type: array
      description:
        example: The mafia member eliminates villagers at night.
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      name:
        example: Mafia
        type: string
      slug:
        example: mafia
        type: string
      team:
        enum:
        - mafia
        - village
        - independent
        example: mafia
        type: string
      video:
        example: https://example.com/video.mp4
        type: string
    type: object
  handler.RoleTemplateResponse:
    description: Role template data
    properties:
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      description:
        example: Standard 8-player game setup
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      name:
        example: 8-player standard
        type: string
      player_count:
        example: 8
        type: integer
      roles:
        items:
          $ref: '#/definitions/handler.TemplateRoleEntryResponse'
        type: array
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
    type: object
  handler.TemplateRoleEntryResponse:
    description: Role entry in a template
    properties:
      count:
        example: 2
        type: integer
      role:
        $ref: '#/definitions/handler.RoleResponse'
      role_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: Mafia Night Support
    url: https://github.com/mafia-night
  description: REST API for the Mafia Night game platform.
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: Mafia Night API
  version: "1.0"
paths:
...
  /admin/role-templates:
    post:
      consumes:
      - application/json
      description: Creates a new role template. Requires admin authentication.
      parameters:
      - description: Template data
        in: body
        name: body
        required: true
        schema:
          properties:
            description:
              type: string
            name:
              type: string
            player_count:
              type: integer
            roles:
              items:
                type: object
              type: array
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.RoleTemplateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
      security:
      - BearerAuth: []
      summary: Create a role template
      tags:
      - admin-role-templates
  /admin/role-templates/{id}:
    delete:
      description: Deletes a role template. Requires admin authentication.
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
      security:
      - BearerAuth: []
      summary: Delete a role template
      tags:
      - admin-role-templates
    patch:
      consumes:
      - application/json
      description: Updates an existing role template. Requires admin authentication.
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        type: string
      - description: Partial template data
        in: body
        name: body
        schema:
          properties:
            description:
              type: string
            name:
              type: string
            player_count:
              type: integer
            roles:
              items:
                type: object
              type: array
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RoleTemplateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
      security:
      - BearerAuth: []
      summary: Update a role template
      tags:
      - admin-role-templates
...
  /role-templates:
    get:
      description: Retrieves all role templates. Optionally filter by player_count.
      parameters:
      - description: Filter by player count
        in: query
        name: player_count
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.RoleTemplateResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
      summary: List role templates
      tags:
      - role-templates
  /role-templates/{id}:
    get:
      description: Retrieves a single role template with its role assignments.
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RoleTemplateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponseBody'
      summary: Get a role template by ID
      tags:
      - role-templates
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"